null catchers added

+++ src/main/java/com/qa/bookings/service/BookingService.java
@@ -8,45 +8,53 @@ import com.qa.bookings.object.Booking;
 import com.qa.bookings.repository.BookingRepo;
 
 @Service
-public class BookingService implements BookingInterface{
-	
+public class BookingService implements BookingInterface {
+
 	private BookingRepo bookRepo;
-	
+
 	public BookingService(BookingRepo bookRepo) {
 		this.bookRepo = bookRepo;
 	}
 
 	@Autowired
 	public JmsTemplate jmsTemplate;
-	
+
 	public String createBooking(Booking booking) {
 		this.bookRepo.save(booking);
 //		jmsTemplate.convertAndSend("Booking List", booking);	
-		return "Booking has been successfully created";
+		return "Booking has been successfully created.";
 	}
-	
+
 	@Override
 	public Booking readBooking(long id) {
-		return bookRepo.findById(id).get();
+		if (bookRepo.existsById(id)) {
+			return bookRepo.findById(id).get();
+		}
+		return null;
+
 	}
 
 	@Override
 	public String updateBooking(long id, String endTime) {
-		Booking aBook = readBooking(id);
-		
-		if (aBook != null) {
+
+		if (bookRepo.existsById(id)) {
+			Booking aBook = readBooking(id);
 			aBook.setEndTime(endTime);
 			bookRepo.save(aBook);
 			return "End time has been changed.";
 		}
-		
+
 		return "Booking doesn't exist.";
 	}
 
 	@Override
 	public String deleteBooking(long id) {
-		bookRepo.deleteById(id);
-		return "Booking deleted.";
+		if (bookRepo.existsById(id)) {
+			bookRepo.deleteById(id);
+			return "Booking deleted.";
+		}
+		return "Booking doesn't exist.";
+
 	}
 
 }
